feldoglozás javítása

+++ app/Http/Controllers/ProcessRawController.php
@@ -79,7 +79,9 @@ class ProcessRawController extends Controller
             if(StoreData::all()->where('websiteid',$websiteid)->where('pizzaid',$pizzaid)->where('pizzasize',$data->size)->count()!=0){
                 //már a dbben van a pizza
                 //dd('már benne van storedatába');
-                RawPizza::all()->where('id',$data['id'])->first()->delete();
+                if(RawPizza::all()->where('id',$data['id'])->first() != null){
+                    RawPizza::all()->where('id',$data['id'])->first()->delete();
+                }
                 return;
             }
